src/EPR.SubmissionMicroservice.API/Middleware/ExceptionHandlingMiddleware.cs
namespace EPR.SubmissionMicroservice.API.Middleware;

public class ExceptionHandlingMiddleware
{
    private readonly RequestDelegate _next;
    private readonly ILogger<ExceptionHandlingMiddleware> _logger;

    public ExceptionHandlingMiddleware(
        RequestDelegate next,
        ILogger<ExceptionHandlingMiddleware> logger)
    {
        _next = next;
        _logger = logger;
    }

    public async Task InvokeAsync(HttpContext context)
    {
        try
        {
            await _next(context);
        }
        catch (Exception e)
        {
            HandleException(context, e);
        }
    }

    private void HandleException(HttpContext httpContext, Exception exception)
    {
        _logger.LogError(exception, "Unhandled exception");
        httpContext.Response.ContentType = "application/json";
        httpContext.Response.StatusCode = StatusCodes.Status500InternalServerError;
    }
}